+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features and improvements
 
+- App: Add option to pause emulator when the window loses focus. (#181)
 - App: Changed background color around screen to black on windowed mode.
 - CD Block: Implement Put Sector command, used by After Burner II. (#78)
 - Core: Performance improvements, especially for ARM builds. (@Wunkolo)